cvs -d :pserver;message_encoding=euc-jp:username@servername:/dir checkout modulename

その後は、とくにオプションに指定をすることなくGUIなCVSクライアントからEUCでメッセージがコミットできます。
※ wincvs1.2ごった煮版同様、漢字変換ライブラリとしてack1.39(小笠原さん作)を使用させていただいています。

2004/08/02 cvsnt-sjis-2.0.51a-1
cvsnt-2.0.51aベースに更新。2.0.14からの変更は本家のリリースノートを参照してください。
※ ただしSJIS版は、デフォルトのままだと2.0.14互換モードになります。
互換モードとは、以下の2.0.14から変更された動作を無効にしたモードです。

1. 置換モードが、最後にファイル追加した時のモードになる機能の無効化
(よい機能にみえるが、ブランチへの変更がトランクへも影響することになってしまう)
2. リポジトリへのファイル名記録の無効化
(これに対応していないGUIなCVSクライアントはcvs logで表示が変になったりする)
3. Attic内外移動の停止機能の無効化
(リポジトリ構成の変更はViewCVS等のリポジトリを直接みるツールへの影響があるかもしれない)
4. ネットワーク共有へのアクセスをエラーにする機能の無効化
(ネットワーク共有上のリポジトリへのアクセスは、危険だといわれているけれども、pserverをたてられない場合に対応するため)

本家と同じ動作をさせるには、CVSROOT/configに

CVSNTSJIS2.0.14CompatibleMode=no

を追加します。

+++ src/expand_path.c
@@ -270,7 +270,7 @@ expand_path (name, file, line)
 char *cvspid()
 {
 	static char tmp[32];
-	sprintf(tmp,"%08x",getpid());
+	sprintf(tmp,"%08x",(int)getpid());
 	return tmp;
 }
